Add admissions and financial aid office picture
