+++ app/src/main/java/com/example/wheaton/First_Year/admissions_office.java
@@ -1,43 +1,42 @@
-package com.example.wheaton;
+package com.example.wheaton.First_Year;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
-import android.net.Uri;
 import android.os.Bundle;
 import android.text.util.Linkify;
-import android.view.Gravity;
 import android.view.View;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
 import android.widget.TextView;
-import android.widget.Toast;
 
-public class financial_aid_office extends AppCompatActivity {
+import com.example.wheaton.R;
+
+public class admissions_office extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_financial_aid_office);
+        setContentView(R.layout.activity_admissions_office);
 
         Intent call_intent = getIntent();
 
         // Phone link
-        TextView phone_num = findViewById(R.id.fin_num);
+        TextView phone_num = findViewById(R.id.ad_num);
         Linkify.addLinks(phone_num, Linkify.PHONE_NUMBERS);
 
         // Make textview clickable
-        TextView website_fin = findViewById(R.id.fin_link);
+        TextView website_fin = findViewById(R.id.ad_link);
         website_fin.setClickable(true);
     }
 
-    public void fin_link(View view){
+    public void ad_link(View view){
         // Set up and call web view
-        WebView webView = (WebView) findViewById(R.id.financial_web_view);
+        WebView webView = (WebView) findViewById(R.id.admissions_web_view);
         webView.setVisibility(View.VISIBLE);
         webView.setWebViewClient(new WebViewClient());
-        webView.loadUrl(getString(R.string.financial_office_url));
+        webView.loadUrl(getString(R.string.admissions_office_url));
 
         WebSettings webSettings = webView.getSettings();
     }
