+++ app/src/main/res/layout/activity_academic_calendar.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".academic_calendar">
+    tools:context=".First_Year.academic_calendar">
 
 
     <WebView
