+++ app/src/main/res/layout/activity_admissions_office.xml
@@ -4,16 +4,16 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".admissions_office">
+    tools:context=".First_Year.admissions_office">
 
     <ImageView
         android:id="@+id/imageView"
         android:layout_width="wrap_content"
-        android:layout_height="210dp"
+        android:layout_height="213dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/registrar_office" />
+        app:srcCompat="@drawable/admissions_financial_aid_office" />
 
     <ScrollView
         android:id="@+id/ad_info"
