+++ app/src/main/res/layout/activity_clubs.xml
@@ -4,4 +4,4 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".clubs" />
+    tools:context=".First_Year.clubs" />
