+++ app/src/main/res/layout/activity_education.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".education">
+    tools:context=".First_Year.education">
 
     <TextView
         android:id="@+id/textView3"
