+++ app/src/main/res/layout/activity_registrar_office.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".registrar_office">
+    tools:context=".First_Year.registrar_office">
 
 
     <ImageView
